requirements.txt
Django==2.2
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
